XP-998 Rename ModuleRelativeResolver to ApplicationRelativeResolver
- Rename ModuleRelativeResolver to ApplicationRelativeResolver
- Move com.enonic.xp.module.ApplicationRelativeResolver to com.enonic.xp.app.ApplicationRelativeResolver
